convert app wrapper to TS
create redux state for IE - error state was
only used for by App and the usage in App was
never actually invoked.

remove previous error redux state as it was
replaced by IE redux state.
