+++ src/interface/actions/internetExplorer.ts
@@ -1,0 +1,6 @@
+export const INTERNET_EXPLORER = 'INTERNET_EXPLORER';
+export function internetExplorer() {
+  return {
+    type: INTERNET_EXPLORER,
+  };
+}
