--- src/interface/selectors/error.ts
@@ -1,3 +1,0 @@
-import { RootState } from 'interface/reducers';
-
-export const getError = (state: RootState) => state.error;
