+++ src/interface/selectors/internetExplorer.ts
@@ -1,0 +1,3 @@
+import { RootState } from 'interface/reducers';
+
+export const isInternetExplorer = (state: RootState) => state.internetExplorer;
